Added first code for managing the diagrams.
The class DiagramHistory implements a continuous view over all diagrams
used in a Toscana session, combining the model behind the protocol, list of
titles in the view and the scale manager from T2/3. It extends AbstractListModel
so it can be displayed easily in a JList from Swing.

The listview for this is DiagramHistoryView which extends JList and implements
a ListCellRenderer as inner class that renders the different parts of the list in
different fonts (currently normal for past diagrams, bold for current diagrams
and italic for forthcoming diagrams).

The view is currently added to the east position in the MainPanel for testing
purposes. The controller part is still lacking.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@22 267469b6-68fb-4578-8a4b-5844dc4013a7