Turned DiagramHistory into a controller.
We moved DiagramHistory into the controller.fca package since it is not
really a model part and it should handle the zooming, too. The next() method
now takes the concept that was used for zooming, the NodeView already
calls this.

For being able to access the same (and only one) instance across the program
we turned it into a Singleton.

Coworker: Peter Becker


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@37 267469b6-68fb-4578-8a4b-5844dc4013a7