Switching diagrams is now possible.
The DiagramHistory supports now switching the diagrams on calls to next(),
which are caused by double-clicks on nodes. The zoomed concept is stored
but not used yet, DiagramView listens to changes on the history and updates
if needed.

To distinguish between different update notifications we added the source of
the change to the ChangeObserver callback.

Some of the changes (the inner class DiagramReference) slipped through into the
last commit -- sorry for that (my fault: Peter).

Coworker: Peter Becker


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@39 267469b6-68fb-4578-8a4b-5844dc4013a7