Some code cleanup.
The former DiagramHistory is now called DiagramController, the parts that
are really storing the history (and implementing the ListModel for Swing) are
in an inner class. This is still not nice OO (the inner class members are
accessed mostly directly), but at least it separates the controller and data
aspects.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@44 267469b6-68fb-4578-8a4b-5844dc4013a7