Only object labels now react on the change of display type.
Now attribute and object labels use different subclasses of LabelView which
is now abstract. The only distinctions made for them is the position (the
placement parameter in the constructor was dropped) and the default setting
for the type (attributes are displayed as list, objects as number).

When changing the type per menu only object labels get the message.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@49 267469b6-68fb-4578-8a4b-5844dc4013a7