Non-realised concepts are now drawn differently.
The concepts are still there, it is just checked if their extend matches the
extent of another concept in the ideal.

If a concept is not realised it is just drawn with 1/3 of the normal radius, the
distinction is handle on DiagramNode (i.e. in the model). We might want to
reconsider this (but maybe not).


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@53 267469b6-68fb-4578-8a4b-5844dc4013a7