Added color gradient.
The nodes now draw themself as black circles with a filling dependent on the
extent size of their concept in relation to the number of objects in the diagram.

The colors are still hard-coded and only one type of gradient is supported. This
should be easy to change.

Note that six different gradients are possible by comparing either the extent or
the contingent of current concept with one of these three:
- number of all objects filtered for the current diagram by extent
- number of all objects filtered for the current diagram by contingent
- all objects in the data set (DB version: SELECT count(*) FROM table;)

The first two could be switched by the normal filter (otherwise it gets too complex
anyway), which would produce a (contingent/extent) x (current/all objects)
matrix of four gradient types.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@61 267469b6-68fb-4578-8a4b-5844dc4013a7