Implemented basic printing support.
ToscanaJ can now print the current diagram, scaled to fit the page format
given. No further print options yet.

Somehow the [Properties] button in the Windows print dialog does not seem
to work within ToscanaJ -- there seems to be no way to give the PageFormat
to the printDialog() and no way to get a pageFormat out of it again. If no
PageFormat is set on the PrinterJob, it will use default settings of Java and not
the default settings of the printer.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@63 267469b6-68fb-4578-8a4b-5844dc4013a7