Removed Windows LnF.
The Windows Look and Feel was introduced too late to be used, to get the
LnF you have to set it before the new components are created (i.e. before the
buildPanel() call). But the LnF is rather ugly: the divider for the splitPane is far
too small to hold the toggle buttons and toolbar and menu bar look not like the
real thing (which is a movin target at the moment anyway).

Probably we should put this as an option into the program, stored with the
session management.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@66 267469b6-68fb-4578-8a4b-5844dc4013a7