Changed handling of model coordinates.
Diagrams will now always use the same coordinate system with y pointing
downwards, this makes handling the diagrams in other parts of the program
easier since more consistent.

This is done by checking and maybe fixing the direction on the first read
access to SimpleLineDiagram and thus the responsibility of the diagram itself,
which is probably A Good Thing (tm).


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@99 267469b6-68fb-4578-8a4b-5844dc4013a7