Refactored the whole layouting code.
The ToscanaGraphics2D class has been completely removed, ToscanaJ now
scales using normal affine transformations on the whole diagram. The only
exception are fonts which are limited to not get below a fixed size (currently
8pt).

This has been done due to serious layouting problems and since the old
approach is rather weird anyway ;-) Nested diagrams now look quite ok, some
details are left to do, but they are close.

All labels end up on the drawing area, although dragging can get weird at the
moment due to the recalculation of the diagram transformation. We have to
stop recalculating the transform during drag events.

Unfortunately Java is not very good in on-screen scaling :-( Sometimes the
lines look very aliased, sometimes fonts leave the labels (which might be my
fault, but maybe not). Check out the SVG export to see how good it could be ;-)

The graphic export shows another advantage of the current scaling approach:
SVGs might turn out to be quite useful to print posters from diagrams that are
not readable anymore on the screen, even printing on a normal page might
turn some diagrams readable. Probably we need some zooming feature on the
screen to place labels etc. for this. Should be easy to do.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@102 267469b6-68fb-4578-8a4b-5844dc4013a7