DB Logger is now determined from properties file.
The log output from the DB connection can now be created/redirected by
changing the properties. Either add the parameter to the 'java' call or enter
something like this to the properties file:

  DBConnection-logger=C\:\\temp\\db.log

Note that you have to escape colons and backslashes. A minus symbol ('-')
puts the output to stdout.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@113 267469b6-68fb-4578-8a4b-5844dc4013a7