Implemented fancy GraphicFormat structure.
The two different writers are now subclasses of the ImageWriter interface,
they define their own GraphicFormats as static inner classes which are the
only way to get the singleton the writers implement.

All writers are initialized by static methods on startup, they register their types
in a GraphicFormatRegistry which handles all the rest. A little unusual approach
but this should be quite stable and rather easy to extend once you understand
the pattern or know how to copy and paste ;-)


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@115 267469b6-68fb-4578-8a4b-5844dc4013a7