Added files to create distribution.
There is now a build script for Ant, start scripts for DOS/Unix, a readme file
and the licence text. All files are in the source module -- it seemed the only
reasonable place. The start scripts are added as binaries to ensure that they
are always using the proper format for their target platform.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@133 267469b6-68fb-4578-8a4b-5844dc4013a7