Added option to add special queries to the CSX files.
In a <database> section a new element <specialQuery> is allowed, which has
to have a unique [name] attribute and has to contain a query part like e.g.
"AVG(column)". The name will be used to create a new entry below the
number/list options in the view menu, in the same group. If this entry is chosen
the object labels will display a query generated from the given query part in
the CSX file.

The additional attribute [format] on <specialQuery> can be used to format the
result using DecimalFormat from the java.text package.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@150 267469b6-68fb-4578-8a4b-5844dc4013a7