New query implementation.
A pretty big and not completely finished change: the binary distinction between
list of objects and the number of objects (plus special queries) is now dropped,
we only have a set of queries, some return only one value (aggregates), others
multiple (lists).

A query is stored as derivate of the abstract base class Query. It has a name
for display in the menu and all of them are collected as part of the schema and
added to the menu, using the first entry as default option. When using in-memory
storage we just use two simple derivates indicating the MemoryMappedConcept
class what to return.

If we use DB queries we use derivates of the class DatabaseInfo.DatabaseQuery,
where we store the SQL queries to do. Multiple columns can be queried at once,
but not yet be displayed. Unless explicitely stated otherwise the two default
queries (list of key names, number of objects) will be added before the user-
defined ones.

The XML works like this:
- in the <database> element we now need <url>, <table> and <key> (<dsn> still
  works but IMO we should drop it)
- if only these are given we end up with the default queries
- if additionally <queries> is given, it contains a list of other queries and it can
  have the attribute "dropDefaults" set to "true" -- in this case the default
  queries will not be created
- in the queries we have either <list> or <aggregate> nodes, each defining a
  specific query. They have the attributes "name" (for the menu), "head" (for
  formatting, see below) and on the list additionally "distinct", which turns the
  SQL option "DISTINCT" on if set to "true"
- in each query there are one or more <column> elements, each defining one
  column in the output (for things like "Price: $5 - $50" queried from two DB
  columns)
- each column has the optional attributes "name" (for later usage in headers),
  "format" (format string) and "separator" (text behind the column) and it contains
  an SQL part to query the appropriate thing (e.g. a column name for a list or
  an SQL aggregate for a number).

The PC-Test example in CVS is updated to show the new structure.

The output should later look like this:
   [head of query][value of column1 using its format][separator of column1]...

TODOs:
- implement formatting/multicolumn results
- fix the calculation which labels are visible and which not
- check if we can do more caching of results

Open questions:
- shouldn't we offer "distinct" for both types?
- the names of the columns don't really match with the string creation -- do we
  want two modes (table or strings) in the long run?


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@168 267469b6-68fb-4578-8a4b-5844dc4013a7