Added more options for DiagramSchema into properties.
All options stored in DiagramSchema are now settable in the properties
file. No UI, no separation from the rest of the properties stuff, but
it works.

The colors can now be given in Hex, too -- similar to HTML, i.e. "#ffffff"
for white, "#ff6600" for some kind of orange, etc.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@180 267469b6-68fb-4578-8a4b-5844dc4013a7