Removed color options from menu.
I don't think we should offer separate color options in the default
UI anyway and the current implementation was nothing more than a demo.

Later we could add a tool for editing schemas and a "Load Schema"
option. The one thing that would make sense within ToscanaJ is changing
the gradient type and reference.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@181 267469b6-68fb-4578-8a4b-5844dc4013a7