Preparing viewer.
We not query the <key> with all list-based queries, so we can get additional
information about objects. Double-clicks already put the key on stdout.

Currently the code gets a bit messy, we need to refactor the whole DB stuff.
I am thinking about moving the query logic completely into the Query subclasses
and encapsulating the results in a new class.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@184 267469b6-68fb-4578-8a4b-5844dc4013a7