Implemented first database viewer.
This is still a bit messy and it does only text-based display, no
formatting, but the infrastructure is now there and proven to work.

The implementation is half-way to a plugin-API, the viewer-specific
things should all be in the DatabaseViewer class, while DatabaseViewerSetup
gives all information needed in structured, but general form (parameters
as String->String Hashtable, template either as JDOM Element or String).
Missing is the dynamic loading and the abstraction into an interface.

I did some simple tests with pctest in my postgres-version, but I think
it should work on other systems, too.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@185 267469b6-68fb-4578-8a4b-5844dc4013a7