Added reports, added exception handling for viewers.
The DatabaseViewers now offer exception handling, errors that happen
during initialization can be passed to the UI. Errors while showing
objects can not be passed, but since something should be displayed
anyway, it seems safe to assume a viewer knows a way to give this
information itself.

An API for generating reports has been added and is called whenever
a double-click on an ObjectLabel displaying an aggregate is detected.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@193 267469b6-68fb-4578-8a4b-5844dc4013a7