Context menus for views/reports, some bugfixes.
The object labels now have context menus if views and/or reports
are defined. The views will be available only if a list is displayed,
reports are always available.

Viewers and report generators do get removed now before a new CSX
is loaded. It is not tried anymore to open a view or report on double-
click if no such thing is defined.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@195 267469b6-68fb-4578-8a4b-5844dc4013a7