Some dbviewer updates.
The HTML viewer and report now use "content" instead of "name" as the
attribute on <field>.

The templates can now have an "url" attribute instead of the content,
the content of the target file will be inlined automatically (works
at the moment only for XML, String will follow).

The HTMLDatabaseViewer uses this. Relative links are possible as usual
(i.e. relative to CSX file).


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@214 267469b6-68fb-4578-8a4b-5844dc4013a7