Added switching content of single labels.
The context menu of the object label views now allows changing the query type
for a single label. All three options (queries, object views, object list
views) are in submenues of the context menu.

The Query class does use a simple registry approach to store information on which
queries are available.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@237 267469b6-68fb-4578-8a4b-5844dc4013a7