Fixed serious bug with filtering.
Whenever a concept had no object clause attached, it would get no filter
clauses neither. If the extent was shown, the object clause would be
calculated from the clauses in the ideal, but since the filter clauses
where not attached, the nodes would display the information for the
unfiltered diagram.

Fixed.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@246 267469b6-68fb-4578-8a4b-5844dc4013a7