Changed highlighting behaviour.
If an inner node is highlighted, all its outer nodes (at most one at the
moment) are highlighted, too. This did involve adding the notion of a parent
node into the inner nodes.

The width of the lines between and around nodes in the highlighted parts can
now be set bz putting something like:

  diagramSchema-selectionLineWidth=7

into the properties file.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@248 267469b6-68fb-4578-8a4b-5844dc4013a7