Additions to AnacondaJ
Adding a substantial number of changes and additions to AnacondaJ.
The most significant is a view model and a view model view. It has been
decided in discussions between myself and Peter Becker that a view model
and a view model view are inferior to a workflow directly encoded via
controllers. These classes are being added though because they have
which can be reworked to serve other purposes.

RJ.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@291 267469b6-68fb-4578-8a4b-5844dc4013a7