Added ConceptInterpretationContext.
A new class stores now all information needed to interpret a concept,
i.e. it gives a context for interpreting concepts. This shall contain
all filtering and nesting concepts, plus the filtering and display
modes (contingent vs. extent).

One of these is now attached to all diagram views and it is already
used for storing the display mode.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@414 267469b6-68fb-4578-8a4b-5844dc4013a7