AnacondaJ now displays object labels.
This is quite a hack, but a small one: if the object labels do not
have a query attached to them, they just show the list of all
objects, which means in the case of AnacondaJ they show just the
SQL fragments attached.

This should probably be refactored either in a way that
we have two types of object labels or somehow that the query
objects are stored somewhere else.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@454 267469b6-68fb-4578-8a4b-5844dc4013a7