Hack to colorize contingent of concepts.
If you add diagramSchema-gradientType=contingent to your toscanaj.prop, the nodes will be colored according to their contingent.

The implementation of this hack has been done by reusing the code of getRelativeExtentSize, changing only some chars. I suppose this lets some refactoring to be done, but I'll let this to someone with a working IDEA (e.g. ;-).


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@478 267469b6-68fb-4578-8a4b-5844dc4013a7