Added some code to have objects directly in the CSX.
After having the concept interpreter, doing two different
versions of CSX -- one memory-based, one DB-mapped -- seems
to be a good idea again. This commit creates the option to
show single diagrams without nesting from those files.

The way the queries are handled is still a hack -- the
object labels still know about the queries, which is not
good.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@502 267469b6-68fb-4578-8a4b-5844dc4013a7