Implemented scaling for Cernato models, refined tests.
The many-valued data from Cernato can now be scaled into formal
contexts for each view, using ViewContext which implements the
Context interface and has an internal implementation of the
interface BinaryRelation.

The tests have been refined to reuse a larger example which
gives a full data set plus the scaled relation to test.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@513 267469b6-68fb-4578-8a4b-5844dc4013a7