Fixed movement of nodes with higher values in a dimension.
The change of the base vector is now relative to the node value in
the dimension, i.e. if a node has a value of 3 in a specific dimension,
the change of the base vector will be only one third of the
usual change, so the node movement matches the mouse movement
again.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@523 267469b6-68fb-4578-8a4b-5844dc4013a7