Added attribute views.
The <views> element in CSX can now contain <attributeView>
elements, which give the option to use the usual database
viewers for attributes. The name of the attribute will be
used as key value, there is no option to have list views.

The two elements <table> and <key> should be given, since
you probably won't have the attributes in the standard
table.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@551 267469b6-68fb-4578-8a4b-5844dc4013a7