Changed scaling behaviour when nesting.
The old scaling approach relied on resizing the labels if they get too small.
This did work using the font size, which did depend on the graphic context,
i.e. it differed for different views (screen, printer, graphic export).

The new approach scales both the inner and the outer diagram in parts.
The scaling of the outer diagram is fixed at the moment (it gets three times
larger), while the inner diagram gets smaller by a calculated amount.

This way the size of the labels always changes by the fixed factor, labels
should not get too small and the hack to avoid too small fonts should be
superflous.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@590 267469b6-68fb-4578-8a4b-5844dc4013a7