Diagram descriptions can now be accessed in nested diagrams.
If a nested diagram is displayed and at least one of the involved simple
diagrams has a description, the button and menu entry for displaying the
description is enabled and a popup menu will ask the user which one to
display. For consistency purposes this also happens if only one diagram
has a description.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@603 267469b6-68fb-4578-8a4b-5844dc4013a7