Node sizes can be scaled according to the contingent or extent sizes.
Similar to the gradient colors the nodes can now change their size according
to their concept's contingent or extent sizes. The reference is always the
largest contingent/extent in the diagram.

The option can be turned on by setting the property 
"diagramSchema-scaleNodeSize" to either "contingent" or "extent", an
explicit "none" or the absence of the property turns it off.

In addition to this the not realised nodes can get a fixed color by using
the property "diagramSchema-notRealisedDiagramNodeColor" to a usual
color value. This way the not realised nodes can be distinguished from the
ones that are small from the scaling.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@605 267469b6-68fb-4578-8a4b-5844dc4013a7