Added animations for temporal component.
The transitions in can now be shown either statically (all at once) or animated. 
The code is flexible in respect with timing, using a ramp model for visibility:
it has a fade-in time, a time of full visibility and a fade-out time. These
values are still fixed in the caller code, but only since we lack the UI for
setting them.

The reference is a fraction of a transition step, each animation element
knows its position, the time for a transition step can be set as milliseconds
value.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@653 267469b6-68fb-4578-8a4b-5844dc4013a7