Added image export for temporal analysis.
A sequence of images can now be exported. A normal file save dialog is
used to get a single file name and location, the complete transition picture
is saved into this file (with its extension determining the format), then a
sequence of transition steps is generated into files with a longer file name,
extended by -000, -001 and so on, where the number of zeroes adjusts
to the length of the largest number.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@665 267469b6-68fb-4578-8a4b-5844dc4013a7