Bugfix: labels now remember what dragging operation is going on.
The label views now remember a dragging state which is calculated on
each drag start, and then kept until a new drag starts. This way we avoid
changing the dragging operation if the mouse pointer moves relative to
the label, as described in bug #554094 (which is only one possible problem
of the old approach).


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@690 267469b6-68fb-4578-8a4b-5844dc4013a7