Added some documentation on installing and configuring ToscanaJ.
This is a DocBook article, which I have not tested yet. I will add tools for
a DocBook -> FOP -> PDF process soon -- once I figured out how to do it.

The idea is to use a build process based on Ant like this:
http://www.dpawson.co.uk/docbook/ant.html

but using both Xalan and FOP from Apache.

Other documentation can be found here:
http://supportweb.cs.bham.ac.uk/documentation/tutorials/docsystem/build/tutorials/docbooksys/segmentedhtml/ch05.html


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@707 267469b6-68fb-4578-8a4b-5844dc4013a7