Added quick hack to support percentages.
This seems to work ok, but it is not yet configurable from the CSX file. A
query for the percentual distribution of the contingent will be added as
one of the default queries, but if the defaults are not used, it will not be
there. I still have to find a way to generalize this notion into the API
properly.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1047 267469b6-68fb-4578-8a4b-5844dc4013a7