Extended percentage queries.
The idea of showing the percentage of objects in a particular concept
contingent/extent is now extended into arbitrary SQL aggregates that
produce numbers. The idea is that a query can be defined like this:

    <distributionQuery name="Average Price (relative)">
      <queryField format="0.00 %">AVG(price)</queryField>
    </distributionQuery>

which means: take the AVG(price) of the concepts contingent/extent
(depending on display mode) and divide it by the AVG(price) of the top
nodes extent (i.e. the object set for the diagram), then display this using
the usual formatting rules.

This can be done using any SQL aggregate suited for division, the normal
case of the object distribution maps down to:

    <distributionQuery name="Object distribution">
      <queryField format="0.00 %">COUNT(*)</queryField>
    </distributionQuery>

Things to see:
- is this useful?
- are there other generalizations of the old percentage display?
- is the tag name ok?
- what happens if the queries return results that are not numbers?

and more testing....


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1052 267469b6-68fb-4578-8a4b-5844dc4013a7