Bugfix/refactoring for handling object label menus.
The object label menu wasn't available on the default count and 
percentage content, which is fixed now for the DB version.

This involved moving the distinction between memory mapped and DB
into the concept interpreters, where it belongs.

Still working on this area, at least the memory based version needs some
more fixing and the first time it is used the label content marker is not
set in the menu. (the radio button thingie)


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1060 267469b6-68fb-4578-8a4b-5844dc4013a7