Added support for transparent background.
By default the diagram is now drawn without background color (on a gray
panel, though). This changes behaviour in printing and diagram export,
where the diagram is used with alpha channels if possible. Only JPEGs
need an explicit background, which will be white now.

This commit changes the semantics for the "as on screen" approach a bit,
but I think it is still reasonable to argue that the diagram itself has no
background color, it is just the gray of the panel (which is actually a
system color and dependend on the Look and Feel).


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1073 267469b6-68fb-4578-8a4b-5844dc4013a7