Bugfix: label size is now calculated on current font.
The label size when drawing was calculated properly on the font set, but
the size for finding the label bounds was based on the default font of
the graphic context, causing problems esp. when the font set was larger
than usual (e.g. when the minimum font size was set). This is now
fixed.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1085 267469b6-68fb-4578-8a4b-5844dc4013a7