Objects can now be added in Siena, some refactoring.
New objects can now be added in Siena. There is still a bit of a redrawing
issue left with this and it seems to break the popup menu for the textual
values, but in general it works.

At the moment this feature is in a simple toolbar on top of the context.
Two other buttons are already added for adding attributes and creating
diagrams, although these functions have not yet been implemented.

Some refactoring has been done aside, e.g. CernatoObject has been
removed in favour of a WritableFCAObjectImplementation and the
ObjectDialog now gets a specific object instead of the whole context
and a position.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1134 267469b6-68fb-4578-8a4b-5844dc4013a7