Reduced number of queries.
The old version of ToscanaJ always queried once for the extent sizes 
and once for the contingent sizes. Since the extent sizes can be 
calculated from the contingent sizes easily, this was a waste of 
resources -- esp. on large DBs.

Now the extent sizes are calculated as sums of the contingent sizes.
This way far less queries have to be done, at most half of the number
done before, sometimes even less if some contingents are not used in a
diagram.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1142 267469b6-68fb-4578-8a4b-5844dc4013a7