Even less count queries.
This time changing the object display mode does not requery anymore.

This means that if there are N contingent clauses in a diagram and it gets
displayed with extents, contingent, even as distribution, we will have
made exactly N queries for it. Thus opening a diagram might take long,
but unless extra information beyond the number is asked for by the user,
everything afterwards will be fast since it comes from the count caches.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1145 267469b6-68fb-4578-8a4b-5844dc4013a7