Made a test run again.
It actually pointed to a bug which I'll fix later. I suspect it won't hurt, but
it should be fixed. The current test works around this bug, in an obvious
fashion (tweaking a counter). This has to be removed once the original
problem is fixed, but I want to finish something else first.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1148 267469b6-68fb-4578-8a4b-5844dc4013a7