Siena now ensures object set consistency across diagrams on save.
This is only intermediate until we have proper integration of the many-
valued context. For now Siena will collect a set of all objects in the
system and check if each diagram includes all objects. If this is not the
case for a particular diagram, the missing objects will be added to the
top, and possibly a new top will be created to get these (in case the
existing top has attributes in its intent).


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1161 267469b6-68fb-4578-8a4b-5844dc4013a7