Added little program to factorize lattices.
It opens a CXT file and splits it into two parts in all possible ways for a
given size of one factor. Then it creates all nested diagrams and lets
the user view it by clicking them. The number of concepts in each part
and in the full nested diagram is shown.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1179 267469b6-68fb-4578-8a4b-5844dc4013a7