Working on a bug
Bug URL: https://sourceforge.net/tracker/index.php?func=detail&aid=693973&group_id=37081&atid=418904
Added comments at the above location.

The problem is that Dialog doesn't get a reference to a parent frame because
at a time of initialising we call JOptionPane.getFrameForComponent which 
doesn't return a frame, IMO it occurs because  the app is not painted yet.

Possible fixes:
- to pass reference to a frame from the main application to DiagramEditingView,
and then to the dialog
- to initialise parent in the dialog, but it seems impossible in java.

For the moment added another constructor to DiagramEditingView allowing
us to pass parent frame to it and using it in Elba. It seems to fix the problem.

Added todo's to remind us to remove unneeded constructor from DiagramEditingView later.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1198 267469b6-68fb-4578-8a4b-5844dc4013a7