Removed second DiagramEditingView constructor.
Nataliya's change to add the parent frame into the constructor seems to
be a reasonable solution. Alternatively we could have initialized the parent
of the context editor later on by using lazy initialization and assuming
that we are fully initialized and added to an container by then. But the
current solution is more explicit, so probably better.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1200 267469b6-68fb-4578-8a4b-5844dc4013a7