Fixed nested diagrams.
Calculation of extent sizes was completely screwed in nested diagrams.
We really need some test cases for this, but the first thing we need to do
is to calculate the lattice product again -- as we did in earlier versions.

At the moment the extent size is still calculated by summing up the
contingent sizes, iterating through the different downsets. This can
be a lot of iteration effort in larger nestings, therefore we cache the
results. They are used on every single redraw of the diagram, so we
really want them fast.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1218 267469b6-68fb-4578-8a4b-5844dc4013a7