Bugfix: nested line diagrams now always have attribute labels for the inner diagram.
If the top node of the outer diagram was not realized in the old versions,
the inner diagram would never show any attribute labels, since the labels
were added only into the inner diagram of the top node, instead of the
inner diagram of the top _realized_ node. This has been changed.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1219 267469b6-68fb-4578-8a4b-5844dc4013a7