Restructured temporal concept analysis controls.
They are not a dialog anymore, but a panel located to the right of the
diagram using the standard splitter setup with one touch resizing.

The controls are now grouped into sections -- both in terms of code
(methods for each of them) and layout (titled borders).

Next step will be renaming the class, I just want to keep this diff.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1269 267469b6-68fb-4578-8a4b-5844dc4013a7