Label scrollbars are now only visible on mouse over.
The labels show their scrollbars only if needed (as before) and if the mouse
is over them (that's new). The idea is that they are distracting, esp. if the
labels are enlarged to shall all content. Normally noone wants to see them
in graphic exports and printouts, which won't happen now.

OTOH we now have the problem that there is a visible clue missing about
which labels show all their contents and which ones only parts. There has
to be something added giving a hint. One option might be a ragged border.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1274 267469b6-68fb-4578-8a4b-5844dc4013a7