Changed scaling code to use elements as objects.
Using tupels as objects was stupid, it obviously (in hindsight) creates
always nominal scales. Now we use one element of the tupels as object
values, the others as attributes. The index for G is given in the method,
the command line can take it as third param (defaults to the first).


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1322 267469b6-68fb-4578-8a4b-5844dc4013a7