Refactored gradient/node size code.
Both use now an Interface to get the values in [0,1] they need. This
interface can be obtained from the concept interpreter by giving one
of three values (for extent, contingent, fixed).

The idea of the refactoring is (a) to avoid duplicated if/then/else 
statements and (b) to allow for other interval sources later on, e.g.
showing deviation from an expected value.

The new single enum is typesafe (as opposed to the two old ones).


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1398 267469b6-68fb-4578-8a4b-5844dc4013a7