Readded try block in DatabaseConnection trying to load DB driver from the classpath.
The reason for this is that DatabaseDriverLoader will look for drivers 
located in a specified directory only. If we have some driver in the
classpath, but not in the drivers directory - we will fail to load it.

So, for now, we first check for drivers in the classpath and then 
with the  DatabaseDriverLoader 

Not sure if this is a good solution though...


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1408 267469b6-68fb-4578-8a4b-5844dc4013a7