Concept interpreters can now be given in the CSX.
The concept interpreter is now part of the conceptual schema and gets
initialized by the schema itself. If the new element <conceptInterpreter>
is given, the class will be loaded dynamically, otherwise the normal setup
determined by database connected or not is used.

The dynamic system uses the following syntax:

  <conceptInterpreter class="....">
    ....
  </conceptInterpreter>

Where the class attribute determines the class of the concept interpreter
to use and this class will be called using a MyClass(Element) constructor,
which gets exactly this element. This way the element's content can be
used to set up the concept interpreter.

A new implementation for the concept interpreter is given for projecting
tuples. The syntax for this is:
 
  <conceptInterpreter class="net.sourceforge.toscanaj.fca.controller.TupleConceptInterpreter">
    <objectColumn>x</objectColumn>
    <objectColumn>y</objectColumn>
    ...
  </conceptInterpreter>

The x,y,... determine the columns to leave after projection, at the moment
we still use spaces as separators for the columns within the tuples. That
might change to tabs.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1460 267469b6-68fb-4578-8a4b-5844dc4013a7