Changed the way we create grayscale and black and white views.
The old version had special derived DiagramSchemas. Not only did that
imply maintaining these methods for each change of the schema, it also
meant that extra drawings like the temporal arrows wouldn't change
with the settings.

The new version creates a wrapper around the Graphics2D object, which
changes the setColor(..) and setPaint(..) methods to adjust the color
into the mode wanted. This is not yet tested with complex paints, just
with color.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1624 267469b6-68fb-4578-8a4b-5844dc4013a7