Attributes in the single-valued context editor can now be moved.
Via context menu the user can chose to move an attribute either in front
of another or right at the end of the list.

Additionally the submenu for both objects and attributes has been changed
to include the two non-sensical options (same item, next item). These
are now shown, but disabled. This way the list is always the full list, which
should make recognition easier.


git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@1632 267469b6-68fb-4578-8a4b-5844dc4013a7